requirements.txt
asgiref==3.7.0
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
common==0.1.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-dotenv==1.4.2
django-filter==23.2
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-import-export==3.2.0
django-jalali==6.0.1
django-jalali-date==1.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
environment==1.0.0
et-xmlfile==1.1.0
idna==3.4
import-export==0.3.1
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
jdatetime==4.1.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
persiantools==3.0.1
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
xlrd==2.0.1
XlsxWriter==3.1.1
xlwt==1.3.0
zipp==3.15.0
django-fsm-log==3.1.0
